[feat] #137 UI 변경 및 순서 정렬 추가, 로직 분리
